Make labels on latency charts more clear
Many people I showed the application to got confused
by the Regional results.

Most of them started to understand what's going on
only after I explained that Regional is a case when
the edge function always runs in Washington instead of
the closest region to the client.

It seems that the labels on the chart could be updated
to remove this confusion and give more value to people
that don't have time to dive deep into the logic of the
app.

Signed-off-by: Piotr Jastrzebski <piotr@chiselstrike.com>